requirements.txt
certifi==2025.7.9
chardet==3.0.4
charset-normalizer==3.4.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
psycopg2==2.9.10
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.1
requests==2.32.4
rfc3986==1.5.0
sniffio==1.3.1
urllib3==2.5.0
